# Replace Marsaglia polar method with Box-muller to generate a normally distributed random number