@fnrizzi 2023-10-31
can you please also include the code you used to generate all the timings?