@dalg24 2023-10-31
> format is wrong , you need to run the `/scripts/apply-clang-format` before pushing the repo

```diff

diff --git a/algorithms/src/Kokkos_Random.hpp b/algorithms/src/Kokkos_Random.hpp
index 0057c79cb..0ee96098c 100644
--- a/algorithms/src/Kokkos_Random.hpp
+++ b/algorithms/src/Kokkos_Random.hpp
@@ -855,9 +855,9 @@ class Random_XorShift64 {
   double normal() {
     constexpr double M_PI2 = 2.0 * Kokkos::numbers::pi_v<double>;
 
-    double u = drand();
-    double v = drand();
-    double r = Kokkos::sqrt(-2.0 * Kokkos::log(u));
+    double u     = drand();
+    double v     = drand();
+    double r     = Kokkos::sqrt(-2.0 * Kokkos::log(u));
     double theta = v * M_PI2;
     return r * Kokkos::cos(theta);
   }
@@ -1099,9 +1099,9 @@ class Random_XorShift1024 {
   double normal() {
     constexpr double M_PI2 = 2.0 * Kokkos::numbers::pi_v<double>;
 
-    double u = drand();
-    double v = drand();
-    double r = Kokkos::sqrt(-2.0 * Kokkos::log(u));
+    double u     = drand();
+    double v     = drand();
+    double r     = Kokkos::sqrt(-2.0 * Kokkos::log(u));
     double theta = v * M_PI2;
     return r * Kokkos::cos(theta);
   }